Adapter, Item 리네이밍
