+++ app/src/main/java/edu/nextstep/camp/calculator/CalculationHistoryItem.kt
@@ -1,0 +1,6 @@
+package edu.nextstep.camp.calculator
+
+data class CalculationHistoryItem(
+    val rawExpression: String,
+    val result: Int
+)
