+++ app/src/main/java/edu/nextstep/camp/calculator/MainActivity.kt
@@ -1,28 +1,24 @@
 package edu.nextstep.camp.calculator
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.isVisible
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import edu.nextstep.camp.calculator.databinding.ActivityMainBinding
-import edu.nextstep.camp.calculator.databinding.ItemResultBinding
 import edu.nextstep.camp.calculator.domain.Operator
 
 class MainActivity : AppCompatActivity(), MainContract.View {
     private lateinit var binding: ActivityMainBinding
     private lateinit var presenter: MainContract.Presenter
-    private val lAdapter by lazy { LAdapter() }
+    private val calculatorHistoryAdapter by lazy { CalculatorHistoryAdapter() }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
         presenter = MainPresenter(this)
+
         initRecyclerView()
 
         binding.button0.setOnClickListener { presenter.addToExpression(0) }
@@ -46,7 +42,7 @@ class MainActivity : AppCompatActivity(), MainContract.View {
 
     private fun initRecyclerView() {
         binding.recyclerView.apply {
-            adapter = lAdapter
+            adapter = calculatorHistoryAdapter
             itemAnimator = null
             layoutManager = LinearLayoutManager(this@MainActivity)
         }
@@ -56,9 +52,9 @@ class MainActivity : AppCompatActivity(), MainContract.View {
         binding.textView.text = rawExpression
     }
 
-    override fun showCalculationHistories(histories: List<LItem>) {
+    override fun openCalculationHistories(histories: List<CalculationHistoryItem>) {
         binding.recyclerView.isVisible = true
-        lAdapter.data = histories
+        calculatorHistoryAdapter.data = histories
     }
 
     override fun closeCalculationHistories() {
@@ -67,37 +63,5 @@ class MainActivity : AppCompatActivity(), MainContract.View {
 
     override fun showCalculationFailMessage() {
         Toast.makeText(this, R.string.incomplete_expression, Toast.LENGTH_SHORT).show()
-    }
-}
-
-data class LItem(
-    val rawExpression: String,
-    val result: Int
-)
-
-class LAdapter : RecyclerView.Adapter<LAdapter.ViewHolder>() {
-
-    var data = emptyList<LItem>()
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        return LayoutInflater.from(parent.context).inflate(R.layout.item_result, parent, false)
-            .let { ViewHolder(it) }
-    }
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.bind(data[position])
-    }
-
-    override fun getItemCount(): Int {
-        return data.size
-    }
-
-    class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-        val binding = ItemResultBinding.bind(view)
-
-        fun bind(item: LItem) {
-            binding.tvExpression.text = item.rawExpression
-            binding.tvResult.text = "= ${item.result}"
-        }
     }
 }
