+++ app/src/main/java/edu/nextstep/camp/calculator/MainPresenter.kt
@@ -47,8 +47,8 @@ class MainPresenter(
         isHistoryDisplayed = !isHistoryDisplayed
         if (isHistoryDisplayed) {
             historySaver.loadHistories()
-                .map { (rawExp, result) -> LItem(rawExp, result) }
-                .also { view.showCalculationHistories(it) }
+                .map { (rawExp, result) -> CalculationHistoryItem(rawExp, result) }
+                .also { view.openCalculationHistories(it) }
         } else {
             view.closeCalculationHistories()
         }
